test: move unit tests for StdoutAction
